Update data_preprocess.py
Corrected IndentationError. In the last FOR loop, "dose_img" was not defined. I also checked the same variable name has been used in another function. Instead of "dose_img", it should be "dose" as per the existing code. To avoid the repetition of the same variable names in multiple instances, I changed "dose" to "doseImg". Though these are minor changes, but still for a newbie it may be hard to solve. I also checked "import argparse" was not mentioned in the script. Hence, added.